+++ student/src/main/java/vn/edu/vnua/fita/student/request/admin/pointofyear/CreatePointYearRequest.java
@@ -25,7 +25,7 @@ public class CreatePointYearRequest {
     @NotNull(message = "Điểm rèn luyện không được để trống")
     @Min(value = 0, message = "Điểm rèn luyện không được dưới 0")
     @Max(value = 100, message = "Điểm rèn luyện không được quá 100")
-    private Integer trainingPoint;
+    private Float trainingPoint;
 
     @NotNull(message = "Tín chỉ tích lũy không được để trống")
     @Min(value = 0, message = "Tín chỉ tích lũy không được dưới 0")
